+++ CHANGELOG.md
@@ -7,6 +7,14 @@ The present file will list all changes made to the project; according to the
 
 ### Added
 - System requirements now check that the security key files (`glpicrypt.key`, `oauth.pem` and `oauth.pub`) can be read, or created when missing.
+- New schemas/endpoints for Service Catalog information in High-Level API v2.4.
+- New schemas/endpoints for reminder translations in High-Level API v2.4.
+- New endpoints for viewing, adding and deleting emails for other users in High-Level API v2.4.
+- Endpoints to add or remove relations for links between assets and appliances in High-Level API v2.4.
+- New schemas/endpoints for custom asset definitions and fields in High-Level API v2.4.
+- New schema/endpoints to view and change Kanban view state data in High-Level API v2.4.
+- `supervisor` property added for `User` schema in High-Level API v2.4.
+- Support for `If-Modified-Since` and `If-Unmodified-Since` HTTP headers for some types of requests in High-Level API. This is not controlled by the API version.
 
 ### Changed
 - Fixed searching values with multiple concurrent spaces.
@@ -41,6 +49,11 @@ The present file will list all changes made to the project; according to the
 ### API changes
 
 #### Added
+- `x-singleton` HLAPI schema property used as a hint for the GraphQL schema generator to create a query that returns a single item, instead of generating a search query.
+- `x-graphql-resolver` HLAPI schema property used as a hint for the GraphQL schema generator to specify a custom resolver for a schema or property.
+  Its use is not necessary unless there are extremely specific, complex requirements for data fetching where it is not possible to use the default resolvers.
+  As a general rule, if you can use the `ResourceAccessor` methods in your REST endpoints, then this property should not be needed.
+- `anyOf` HLAPI schema property used to indicate that a property can be one of multiple types. This is used in combination with the `x-graphql-resolver` property to specify a custom resolver that can return different types of data based on the context of the request.
 
 #### Changes